refactor: move truncate filter file
