+++ frontend/app/src/components/accounts/AccountGroupHeader.vue
@@ -2,7 +2,6 @@
 import { type Balance, type BigNumber } from '@rotki/common';
 import { type ComputedRef, type PropType } from 'vue';
 import Fragment from '@/components/helper/Fragment';
-import { truncateAddress, truncationPoints } from '@/filters';
 import { type XpubAccountWithBalance } from '@/types/blockchain/accounts';
 
 const { t } = useI18n();
