+++ frontend/app/src/main.ts
@@ -1,7 +1,6 @@
 /* istanbul ignore file */ import { PiniaVuePlugin, createPinia } from 'pinia';
 import Vue, { provide } from 'vue';
 import App from '@/App.vue';
-import '@/filters';
 import '@/main.scss';
 import 'roboto-fontface/css/roboto/roboto-fontface.css';
 import 'typeface-roboto-mono';
