+++ frontend/app/src/utils/truncate.ts
